% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSenMC-package.R
\docType{package}
\name{BayesSenMC-package}
\alias{BayesSenMC-package}
\alias{BayesSenMC}
\title{The 'BayesSenMC' package.}
\description{
This package generates different posterior distributions of adjusted odds ratio under different priors of sensitivity and specificity, and plots the models for comparison. It also provides estimations for the specifications of the models using diagnostics of exposure status with a non-linear mixed effects model.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
