% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{drawModels}
\alias{drawModels}
\title{Draw Models}
\usage{
drawModels(modelList, graphNames, a, N1, c, N0, se, sp, lowerprob = 0.01,
  upperprob = 0.99)
}
\arguments{
\item{graphNames}{a vector of strings of names given to the plots, must be of the same length as "modelList".}

\item{a}{# of exposed subjects in the case group. Along with N1, c, N0, se and sp, they are used to plot the solid and dotted line that each represent a model with no misclassification and misclassification with constant Se and Sp.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{se}{sensitivity}

\item{sp}{specificity}

\item{lowerprob}{the lowest value of all the lowerprob-percentiles of all the models that will be plotted will be taken as the lower limit for all the plots. Default to 0.01.}

\item{upperprob}{the highest value of all the upperprob-percentiles of all the models that will be plotted will be taken as the upper limit for all the plots. Default to 0.99.}

\item{modellist}{a list of stanfit objects that are returned by stan() command.}
}
\value{
It creates a models.pdf file under the working directory, that consists of the plots
of the models arranged in an appropriate grid.
}
\description{
This function plots any of the above six models and a solid and dotted line that are the probability density of crude and corrected OR, respectively (more information in the paper of Chu et al. 2006). A file named "draws.pdf" will be returned to your local directory.
}
\examples{
A <- correctedOR(a = 126, N1 = 218, c = 71, N0 = 295)
B <- crudeOR(a = 126, N1 = 218, c = 71, N0 = 295, se = 0.94, sp = 0.97)
C <- logitOR(a = 126, N1 = 218, c = 71, N0 = 295, mx = 1.9814, my = 2.8586,
varx = 0.9872, vary = 0.5806, x = 2.76, y = 3.58)
drawModels(modelList = list(A, B, C), graphNames = c("(A)", "(B)", "(C)"),
a = 126, N1 = 218, c = 71, N0 = 295, se = 0.94, sp = 0.97)
}
