% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fixedCorrOR}
\alias{fixedCorrOR}
\title{Model that extends from logitModel but allows there to be a fixed correlation between Sentivity and Specificity}
\usage{
fixedCorrOR(a, N1, c, N0, mx0, mx1, varx0, varx1, x0, x1, rhose,
  name = "Model with fixed correlation")
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{mx0}{normal distribution of X0 with (mx0, varx0)}

\item{mx1}{conditional distribution of X1 with (mx1, varx1)}

\item{varx0}{normal distribution of X0 with (mx0, varx0)}

\item{varx1}{conditional normal distribution of X1 with (mx1, varx1)}

\item{x0}{used as an initial value of X0}

\item{x1}{used as an initial value of X1}

\item{rhose}{correlation between Se and Sp}

\item{name}{a string of the name of the model. Default to "Model with fixed correlation".}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See http://mc-stan.org/rstan/reference/stanfit-class.html for more details.
}
\description{
This function generates a posterior distribution of corrected odds ratio given nondifferential misclassification with Se=0.5+0.5*expit(X) and Sp=0.5+0.5*expit(Y), where X and Y are independently distributed, and Se and Sp are correlated
}
\examples{
fixedCorrOR(a = 126, N1 = 218, c = 71, N0 = 295, mx0 = 1.9814, mx1 = 2.8586,
varx0 = 0.9872, varx1 = 0.5806, x0 = 1.9814, x1 = 2.8586, rhose = -0.6271)
}
