% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{logitOR}
\alias{logitOR}
\title{Model with nondifferential misclassification with Sentivity and Specificity each under a logit-transformed scaled bivariate normal distribution}
\usage{
logitOR(a, N1, c, N0, mx, my, varx, vary, x, y, name = "Logit Model")
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{mx}{normal distribution of X with (mX, varX).}

\item{my}{normal distribution of Y with (mY, varY).}

\item{varx}{normal distribution of X with (mX, varX).}

\item{vary}{normal distribution of Y with (mY, varY).}

\item{x}{used as an initial value of X.}

\item{y}{used as an initial value of Y.}

\item{name}{a string of the name of the model. Default to "Logit Model".}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See http://mc-stan.org/rstan/reference/stanfit-class.html for more details.
}
\description{
This function generates a posterior distribution of corrected odds ratio given nondifferential misclassification with Se=0.5+0.5*expit(X) and Sp=0.5+0.5*expit(Y), where X and Y are independently distributed.
}
\examples{
logitOR(a = 126, N1 = 218, c = 71, N0 = 295, mx = 1.9814, my = 2.8586,
varx = 0.9872, vary = 0.5806, x = 2.76, y = 3.58)
}
