% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmeNDiff.R
\name{paramEst}
\alias{paramEst}
\title{Model without misclassification}
\usage{
paramEst(model)
}
\arguments{
\item{model}{a GLMM model built with the \code{nlme_nondiff()} function.}
}
\value{
It returns a list of parameter estimates which can be input into the Bayesian model functions in
\code{BayesSenMC}. \code{(mean_logSe, var_logSe)} and \code{(mean_logSp, var_logSp)} are the logit prior distributions for Se and Sp.
\code{Se} and \code{Sp} are the corresponding mean values given the logit prior means. \rho is the correlation estimate between Se and
Sp. \code{fisher_mean} is the Fisher's mean of the correlation assume a Fisher's distribution.
}
\description{
Generate a stanfit object corresponding to a posterior distribution of uncorrected odds ratio given no misclassification.
}
\examples{
mod <- nlme_nondiff(raw) # see nlme_nondiff() for detailed example.
pList <- paramEst(mod)
}
