% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{correctedOR}
\alias{correctedOR}
\title{Model with no misclassification}
\usage{
correctedOR(a, N1, c, N0, name = "Corrected Model")
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{name}{a string of the name of the model. Default to "Corrected Model".}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See
http://mc-stan.org/rstan/reference/stanfit-class.html for more details.
}
\description{
This function generates a posterior distribution of uncorrected odds ratio given no misclassification.
}
\examples{
correctedOR(a = 126, N1 = 218, c = 71, N0 = 295)
}
