% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{crudeOR}
\alias{crudeOR}
\title{Model with nondifferential classification with constant Sentivity and Specificity}
\usage{
crudeOR(a, N1, c, N0, se, sp, name = "Crude Model")
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{se}{sensitivity}

\item{sp}{specificity}

\item{name}{a string of the name of the model. Default to "Crude Model".}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See http://mc-stan.org/rstan/reference/stanfit-class.html for more details.
}
\description{
This function generates a posterior distribution of corrected odds ratio given nondifferential misclassification with constant Se and Sp.
}
\examples{
crudeOR(a = 126, N1 = 218, c = 71, N0 = 295, se = 0.94, sp = 0.97)
}
