% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fixedCorrOR}
\alias{fixedCorrOR}
\title{Model with nondifferential, correlated misclassification}
\usage{
fixedCorrOR(a, N1, c, N0, mx0, mx1, varx0, varx1, x0, x1, rhose,
  name = "Model with fixed correlation", traceplot = FALSE,
  inc_warmup = FALSE, window = NULL, ...)
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{mx0}{normal distribution of X0 with (mx0, varx0).}

\item{mx1}{conditional distribution of X1 with (mx1, varx1).}

\item{varx0}{normal distribution of X0 with (mx0, varx0).}

\item{varx1}{conditional distribution of X1 with (mx1, varx1)}

\item{x0}{used as an initial value for X0}

\item{x1}{used as an initial value for X1}

\item{rhose}{correlation between Se and Sp}

\item{name}{a string of the name of the model. Default to "Model with fixed correlation".}

\item{traceplot}{Logical, defaulting to \code{FALSE}. If \code{TRUE} it will draw the 
\code{\href{https://mc-stan.org/rstan/reference/stanfit-method-traceplot.html}{traceplot}} corresponding to one or more Markov chains.}

\item{inc_warmup}{Only evaluated when \code{traceplot = TRUE}. \code{TRUE} or \code{FALSE}, indicating whether or not to include the warmup sample in the
traceplot; defaults to \code{FALSE}.}

\item{window}{Only evaluated when \code{traceplot = TRUE}. A vector of length 2. Iterations between \code{window[1]} and \code{window[2]} will be shown in the plot. 
The default shows all iterations if \code{inc_warmup} is \code{TRUE} and all iterations from the sampling period only if \code{inc_warmup} is \code{FALSE}. 
If \code{inc_warmup} is \code{FALSE} the iterations specified in \code{window} do not include iterations from the warmup period.
The default number of iterations is 2000 unless otherwise specified in the optional \code{iter} argument.}

\item{...}{optional parameters passed to \code{\href{https://www.rdocumentation.org/packages/rstan/versions/2.17.3/topics/stan}{stan}}.}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See
\code{\href{https://mc-stan.org/rstan/reference/stanfit-class.html}{here}} for more details.
}
\description{
Generate a stanfit object corresponding to a posterior distribution of corrected odds ratio given nondifferential misclassification that extends from the logit model but allows there to be a fixed correlation between Sentivity and Specificity.
}
\examples{
fixedCorrOR(a = 126, N1 = 218, c = 71, N0 = 295, mx0 = 1.9814, mx1 = 2.8586,
  varx0 = 0.9872, varx1 = 0.5806, x0 = 1.9814, x1 = 2.8586, rhose = -0.6271, chains = 3, iter = 10000)

fixedCorrOR(a = 126, N1 = 218, c = 71, N0 = 295, mx0 = 1.9814, mx1 = 2.8586,
  varx0 = 0.9872, varx1 = 0.5806, x0 = 1.9814, x1 = 2.8586, rhose = -0.6271, traceplot = TRUE, window = c(1, 1000))
}
