% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nonDiffOR}
\alias{nonDiffOR}
\title{Model with differential misclassification}
\usage{
nonDiffOR(a, N1, c, N0, mu, varz, z = NULL,
  name = "Model with differential classification", chains = 2,
  traceplot = FALSE, inc_warmup = FALSE, window = NULL,
  refresh = 0, seed = 0, ...)
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{mu}{vector of length 4; multivariate normal distribution of \eqn{z \sim (mu, varz)}, where each \eqn{\mu} corresponds to the logit mean of \eqn{Se_0$}, \eqn{Se_1}, \eqn{Sp_0} and \eqn{Sp_1}.}

\item{varz}{variance-covariance matrix of z, passed in as a 4 by 4 matrix.}

\item{z}{vector of length 4; used as an initial value for \eqn{z \sim (mu, varz)}. Default to mu.}

\item{name}{a string of the name of the model. Default to "Model with differential misclassification".}

\item{chains}{number of Markov Chains. Default to 2.}

\item{traceplot}{Logical, defaulting to \code{FALSE}. If \code{TRUE} it will draw the 
\href{https://mc-stan.org/rstan/reference/stanfit-method-traceplot.html}{traceplot} corresponding to one or more Markov chains.}

\item{inc_warmup}{Only evaluated when \code{traceplot = TRUE}. \code{TRUE} or \code{FALSE}, indicating whether or not to include the warmup sample in the
traceplot; defaults to \code{FALSE}.}

\item{window}{Only evaluated when \code{traceplot = TRUE}. A vector of length 2. Iterations between \code{window[1]} and \code{window[2]} will be shown in the plot. 
The default shows all iterations if \code{inc_warmup} is \code{TRUE} and all iterations from the sampling period only if \code{inc_warmup} is \code{FALSE}. 
If \code{inc_warmup} is \code{FALSE} the iterations specified in \code{window} do not include iterations from the warmup period.
The default number of iterations is 2000 unless otherwise specified in the optional \code{iter} argument.}

\item{refresh}{an integer value used to control how often the progress of sampling is reported. By default, the progress indicator is turned off, thus refresh <= 0.
If on, refresh = max(iter/10, 1) is generally recommended.}

\item{seed}{the seed for random number generation. See \href{https://www.rdocumentation.org/packages/rstan/versions/2.17.3/topics/stan}{stan} for more details.}

\item{...}{optional parameters passed to \href{https://www.rdocumentation.org/packages/rstan/versions/2.17.3/topics/stan}{stan}.}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See
\href{https://mc-stan.org/rstan/reference/stanfit-class.html}{here} for more details.
}
\description{
Generate a stanfit object corresponding to a posterior distribution of corrected odds ratio given a four-variate differential misclassification.
}
\examples{
# Case-control study data of Bipolar Disorder with rheumatoid arthritis (Farhi et al. 2016)
# Data from \\url{https://www.sciencedirect.com/science/article/pii/S0165032715303864#bib13}

nonDiffOR(a = 126, N1 = 218, c = 71, N0 = 295, mu = c(2.76, 2.76, 3.58, 3.58),
  varz = c(0.9100002, 0.8190002, -0.5053286, -0.5053286, 0.8190002, 0.9100002,
  -0.5053286, -0.5053286, -0.5053286, -0.5053286, 0.7300000, 0.6570000,
  -0.5053286, -0.5053286, 0.6570000, 0.7300000), z = c(2.76, 2.76, 3.58, 3.58),
   chains = 3, iter = 10000, seed = 0)

nonDiffOR(a = 126, N1 = 218, c = 71, N0 = 295, mu = c(2.76, 2.76, 3.58, 3.58),
  varz = matrix(c(0.9100002, 0.8190002, -0.5053286, -0.5053286, 0.8190002, 0.9100002,
  -0.5053286, -0.5053286, -0.5053286, -0.5053286, 0.7300000, 0.6570000,
  -0.5053286, -0.5053286, 0.6570000, 0.7300000), 4, 4), z = c(2.76, 2.76, 3.58, 3.58), 
  traceplot = TRUE)
}
