% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nonDiffOR}
\alias{nonDiffOR}
\title{Model with differential misclassification with Se0, Se1, Sp0, and Sp1 to be four-variate logit-transformed rescaled normal distributions}
\usage{
nonDiffOR(a, N1, c, N0, mu, varz, z,
  name = "Model with differential classification")
}
\arguments{
\item{a}{# of exposed subjects in the case group.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{mu}{vector of length 4; multivariate normal distribution of z ~ (mu, varz).}

\item{varz}{covariance matrix of length 4*4, passed in as an one dimentional array of length 16; multivariate normal distribution of z ~ (mu, varz).}

\item{z}{vector of length 4; used as an initial value for Z.}

\item{name}{a string of the name of the model. Default to "Model with differential misclassification".}

\item{mx0}{normal distribution of X0 with (mx0, varx0).}

\item{mx1}{conditional distribution of X1 with (mx1, varx1).}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See http://mc-stan.org/rstan/reference/stanfit-class.html for more details.
}
\description{
This function generates a posterior distribution of corrected odds ratio given nondifferential misclassification with Se=0.5+0.5*expit(X) and Sp=0.5+0.5*expit(Y), where X and Y are independently distributed, and Se and Sp are correlated and the correlation is a random variable, which is equal to 0.5*[In(1+Z)-In(1-Z)] with a normally distributed Z.
}
\examples{
nonDiffOR(a = 126, N1 = 218, c = 71, N0 = 295, mu = c(2.76, 2.76, 3.58, 3.58),
varz = c(0.9100002, 0.8190002, -0.5053286, -0.5053286, 0.8190002, 0.9100002,
-0.5053286, -0.5053286, -0.5053286, -0.5053286, 0.7300000, 0.6570000,
-0.5053286, -0.5053286, 0.6570000, 0.7300000), z = c(2.76, 2.76, 3.58, 3.58))
}
