% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModels.R
\name{OR_hist}
\alias{OR_hist}
\title{Plot Model}
\usage{
OR_hist(model, a, N1, c, N0, se = 1, sp = 1, binwidth = 0.25,
  fill = "gray", ...)
}
\arguments{
\item{model}{A stanfit object.}

\item{a}{# of exposed subjects in the case group. Along with N1, c, N0, se and sp, they are used to plot probability density with no misclassification and constant misclassification as a comparison.}

\item{N1}{# of total subjects in the case group.}

\item{c}{# of exposed subjects in the control group.}

\item{N0}{# of total subjects in the control group.}

\item{se}{sensitivity. Default to 1. If no other values are specified for either se or sp, then only the density curve of corrected model will be drawn.}

\item{sp}{specificity. Default to 1.}

\item{binwidth}{default to \code{0.25}}

\item{fill}{default to \code{"gray"}}

\item{...}{optional additional arguments passed to \code{geom_histogram}}
}
\value{
It returns a \code{\href{https://www.rdocumentation.org/packages/ggplot2/versions/0.9.0/topics/ggplot}{ggplot}} that can be further 
customized using the ggplot2 package.
}
\description{
Plot the posterior distribution of adjusted odds ratio given the stanfit object.
}
\examples{
my_model <- crudeOR(a = 126, N1 = 218, c = 71, N0 = 295, se = 0.94, sp = 0.97, chains = 5, iter = 20000, warmup = 2000)
my_hist <- OR_hist(my_model, a = 126, N1 = 218, c = 71, N0 = 295, se = 0.94, sp = 0.97)
}
