% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{randCorrOR}
\alias{randCorrOR}
\title{Model with nondifferential, randomly correlated misclassification}
\usage{
randCorrOR(
  a,
  N1,
  c,
  N0,
  prior_list = NULL,
  m.lg.se = NULL,
  m.lg.sp = NULL,
  s.lg.se = NULL,
  s.lg.sp = NULL,
  lg.se = NULL,
  lg.sp = NULL,
  m.z = NULL,
  s.z = NULL,
  z = NULL,
  logitpi0_prior = c(0, 10),
  lor_prior = c(0, 2),
  chains = 2,
  traceplot = FALSE,
  inc_warmup = FALSE,
  window = NULL,
  refresh = 0,
  seed = NA,
  ...
)
}
\arguments{
\item{a}{number of exposed subjects in the case group.}

\item{N1}{number of total subjects in the case group.}

\item{c}{number of exposed subjects in the control group.}

\item{N0}{number of total subjects in the control group.}

\item{prior_list}{list of priors. Can be replaced by the function call to \code{paramEst}, or a list of prior parameters
(\code{m.lg.se}, \code{s.lg.se}, \code{m.lg.sp}, \code{s.lg.sp}, \code{m.z}, \code{s.z}).
If \code{prior_list} is specified, the values for the corresponding function parameters will be disregarded.}

\item{m.lg.se}{normal distribution of logit Se with (mean = m.lg.se, sd = s.lg.se). Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{m.lg.sp}{conditional normal distribution of logit Sp given Se with (m.lg.sp, s.lg.sp). Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{s.lg.se}{standard deviation of logit Se. Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{s.lg.sp}{standard deviation of logit Sp. Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{lg.se}{used as an initial value for logit Se. Default to m.lg.se. Do not have to specify this if \code{prior_list} is given - it will be disregarded. Default to \code{m.lg.se}}

\item{lg.sp}{used as an initial value for logit Sp. Default to m.lg.sp. Do not have to specify this if \code{prior_list} is given - it will be disregarded. Default to \code{m.lg.sp}}

\item{m.z}{normal distribution of Z with (mean = m.z, sd = s.z). Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{s.z}{normal distribution of Z with (mean = m.z, sd = s.z). Do not have to specify this if \code{prior_list} is given - it will be disregarded.}

\item{z}{used as an initial value of Fisher's Z transformed of rho, where correlation rho = (exp(2\emph{z)-1)/(1+exp(2}z))). Do not have to specify this if \code{prior_list} is given - it will be disregarded. Default to \code{m.z}}

\item{logitpi0_prior}{mean and sd of the prior normal distribution of \code{logit(pi0)}. Default to \code{c(0,10)}.}

\item{lor_prior}{mean and sd of the prior normal distribution of corrected log odds ratio. Default to \code{c(0,2)}.}

\item{chains}{number of Markov Chains. Default to 2.}

\item{traceplot}{Logical, defaulting to \code{FALSE}. If \code{TRUE} it will draw the \link[rstan]{traceplot} corresponding to one or more Markov chains.}

\item{inc_warmup}{Only evaluated when \code{traceplot = TRUE}. \code{TRUE} or \code{FALSE}, indicating whether or not to include the warmup sample in the
traceplot; defaults to \code{FALSE}.}

\item{window}{Only evaluated when \code{traceplot = TRUE}. A vector of length 2. Iterations between \code{window[1]} and \code{window[2]} will be shown in the plot.
The default shows all iterations if \code{inc_warmup} is \code{TRUE} and all iterations from the sampling period only if \code{inc_warmup} is \code{FALSE}.
If \code{inc_warmup} is \code{FALSE} the iterations specified in \code{window} do not include iterations from the warmup period.
The default number of iterations is 2000 unless otherwise specified in the optional \code{iter} argument.}

\item{refresh}{an integer value used to control how often the progress of sampling is reported. By default, the progress indicator is turned off, thus refresh <= 0.
If on, refresh = max(iter/10, 1) is generally recommended.}

\item{seed}{the seed for random number generation. See \link[rstan]{stan} for more details.}

\item{...}{optional parameters passed to \link[rstan]{stan}.}
}
\value{
It returns a stanfit object of this model, which inherits stanfit class methods. See \link[rstan]{rstan} for more details.
}
\description{
Generate a stanfit object corresponding to a posterior distribution of corrected odds ratio given nondifferential misclassification that extends from the logit model but allows a random correlation between Sensitivity and Specificity.
}
\examples{
# Case-control study data of Bipolar Disorder with rheumatoid arthritis (Farhi et al. 2016)
# Data from \url{https://www.sciencedirect.com/science/article/pii/S0165032715303864#bib13}

mod <- nlmeNDiff(bd_meta, lower = 0) # see \code{nlmeNDiff()} for detailed example.
prior_list <- paramEst(mod)
randCorrOR(a = 66, N1 = 11782, c = 243, N0 = 57973, prior_list = prior_list,
chains = 3, iter = 10000)
}
